Adds preliminary threading tests.
Currently, we've only replaced the HardwareMap object with a slightly
different underlying object - we've gone from a thin wrapper around a
SQLAlchemy Session() to a scoped_session() object. This is basically
just a proxy that creates and maintains a thread-local Session object on
demand - which happens to be exactly what we want.

There are probably other threading violations, but this is a definite
step along the path to coexisting peacefully with threaded code.