py_get_samples: Use IGMPv3 source-specific multicasting (SSM)
This change should have no impact on deployments with a single CRS
board, but reduces link congestion when multiple boards are present on
the same network.

In particular, SSM allows switches to select multicast traffic from a
specific source when requested, rather than directing all multicast
traffic associated with a particular multicast group.

This commit is the important bits from #5. The rest of the PR is an API
improvement work in progress and can be shelved.