regression tests: shift from plain tox to uv / tox-uv
The uv project is much faster, and avoids the need to bootstrap a Python
environment in order to invoke Python tooling to create another Python
environments. (Or if it does, it's behind the scenes now).

The test.sh script tells you how to install uv (rather than doing it for
you - which is a curl | sh invocation that should upset the security
conscious.) This will have to be folded into any github actions that
perform testing.