py_get_samples: sock.settimeout() is incompatible with non-blocking sockets.
Instead, use asyncio.wait_for() to establish a maximum time for recv()
to work. This correctly interleaves recv() calls that were previously
serialized.

This py_get_samples is still not as performant as it should be - I
believe we're no longer hitting some kind of implicit serialization, and
instead, may be hitting either the interpreter's performance limits, or
the syscall overhead associated with the one-packet-per-recv() network
path.