py_get_samples: Adds internal tests for magnitude.
These are currently limited (they do DC magnitude checks only) and this
seems simultaneously the right place to keep them (next to the code to
be tested) and the wrong place to keep them (somewhere outside the test/
directory). This is a "perfect is the enemy of good" commit - better
having them here than nowhere.