py_get_samples: (limited) packet reordering to compensate for out-of-order arrival
We stream via UDP, which makes no guarantees about packet
order-of-arrival - the network could have (but generally didn't) caused
packets leaving the CRS to arrive in a different order. Because this
didn't happen, we didn't bother worrying about it.

Now that the CRS streams much higher data rates, we're engaging more
than 1 CPU core with packet transmission - which also causes packets to
leave the board in a slightly disordered state.

This sorting stage is a *very* limited recovery mechanism. Specifically,
it will help with internal ordering problems - for example, the
following packet stream

    ..., 1, 2 -> [3, 5, 4, 6] -> 7, 8, ...
                     ^--^

It will stumble over boundary ordering problems - for example,

    1, 2, 4 -> [3, 5, 6] -> 7, 8, ..., or
          ^-----^

    1, 2, 3 -> [4, 5, 7] -> 6, 8, ...
                      ^-----^

To manage boundary conditions, it's necessary to use a sliding window -
which is out-of-scope for py_get_samples and properly belongs somewhere
common (like rfmux/streamer.py).