Create noise spectra (#33)
Adds the capability to include noise spectra in the detector digest plots.
For now this includes only the ability to take data from the "slow" streamer, and doesn't yet stitch together the spectra from get_pfb_samples, but will allow users to plot the 1/f and noise spectra for all of their biased detectors, as well as show the GR roll-off for detectors where that is visible within decimation 0.

This feature branch will remain open and be used to integrate the faster spectra as well.