mock: Correct server/client-side split
On embedded deployments, the split is clear (there's a server-side
object and a client-side object, and they can't access each other except
through Tuber RPC calls).

Here, we used to have only a single object that played double duty - it
was exposed via a "fake" Tuber server, and was inconsistently
serialized.

This commit restores the client/server split and removes a number of
serialization workarounds.