c++ packet refactor: correct noise sample generation
Because udp_streamer synthesizes "fake" packets, it needs to be updated
to match c++ packet normalization and indexing. (The on-the-wire format
and normalization has not changed - this is pure API detail inside
Python.)

For example, for long readout packets, packets were previously
internally structured as a 2048-length array of (i, q) pairs. Now, it's
a std::vector<std::complex> of length 1024. Hence, the noise synthesis
code does not need to descend all the way to serialization format when
creating new packets from scratch.