Deprecate conda support; remove mentions in documentation
Conda apparently has broken PyQT6 support, which is why we [gui] was
listed as an optional dependency in pyproject.toml. However, for all
non-conda platforms, that means "pip install -e .[gui]" or equivalent
was mandatory and easy to forget.

Conda seems to be losing ground to uv and other environments, so in
practice, using [gui] as an optional dependency broke more than it
fixed.

If Conda support is still necessary, I suspect there is an in-Conda way
of skipping dependencies and managing them manually.